+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 3.4.1
+## 3.4.2
 
 * Fixes an exception caused by the `onUrlChange` callback passing a null `NSUrl`.
+
+## 3.4.1
+
+* Fixes internal type conversion error.
+* Adds internal unknown enum values to handle api updates.
 
 ## 3.4.0